Frontend dir rename
