+++ app/frontend/css/admin/active_admin_theme.scss
