+++ app/frontend/css/admin/has_many_layout.scss
