+++ app/frontend/entrypoints/active_admin.js
@@ -5,5 +5,5 @@ import "jquery-ui-dist/jquery-ui";
 import "@fortawesome/fontawesome-free/css/all.min.css";
 import "@activeadmin/activeadmin";
 
-import "~/src/chord-forms";
-import "~/src/song-chords";
+import "~/js/chord-forms";
+import "~/js/song-chords";
