+++ app/frontend/entrypoints/active_admin.scss
@@ -1,4 +1,4 @@
 @import "@activeadmin/activeadmin/src/scss/mixins";
 @import "@activeadmin/activeadmin/src/scss/base";
-@import "../styles/admin/active_admin_theme";
-@import "../styles/admin/has_many_layout";
+@import "../css/admin/active_admin_theme";
+@import "../css/admin/has_many_layout";
